Solved problems with moveTo and battle states

+++ client/v3/action/MoveTo.hpp
@@ -9,7 +9,7 @@ class MoveTo : public Action {
  public:
   ~MoveTo();
 
-  static MoveTo* Create(int milliseconds, int x, int y);
+  static MoveTo* Create(int milliseconds, qreal x, qreal y);
   static MoveTo* Create(int milliseconds, QPointF point);
 
   void Step(unsigned int ellapsed) override;
@@ -20,10 +20,10 @@ class MoveTo : public Action {
  private:
   bool done_;
 
-  int delta_x_;
-  int delta_y_;
-  int end_x_;
-  int end_y_;
+  qreal delta_x_;
+  qreal delta_y_;
+  qreal end_x_;
+  qreal end_y_;
   int milliseconds_;
   int timeout_;
   int ellapsed_;
